requirements.txt
asgiref==3.2.10
Django==3.0.8
django-filter==2.3.0
django-redis-cache==2.1.1
djangorestframework==3.11.0
Markdown==3.2.2
psycopg2==2.8.5
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1

